State the material a derived glTF means, instead of inheriting metal
A primitive with no material takes glTF's default material, and that default is
metallicFactor 1.0 — a fully rough metal, which has no diffuse response and
draws as a reflection of its surroundings rather than as its own surface. Every
gltf rendition this converter produced said exactly that, conformantly, so every
correct renderer drew metal and none of them was wrong to. Reported from the
Homeworldz client, which imported metallic 1.00 and photographed three texture
plates rendering as flat sky-coloured quads while their textures decoded
perfectly.

The converter is the right place to fix it. The content is a mesh a person
uploaded through a viewer, which is not metal, and the alternative is each
client guessing and disagreeing about what the asset failed to say. It now
writes metallicFactor 0 and roughnessFactor 1 and points the primitive at it.
No baseColorTexture: a type-49 mesh carries geometry and its faces are textured
by the region's texture entry, so naming an image here would invent one.

Uploaded GLBs are deliberately untouched. Their materials are the creator's
statement, and rewriting one on ingest would be wrong for anybody who genuinely
uploads a metal. The three plates in the report are uploads, not renditions, so
this fix does not reach them — their omission was authored upstream.

The generator moves to 0.8 so regeneration reissues the renditions that drew as
metal.

Also casts the two int8_t fields in the inventory refusal log. Streaming one
writes the character with that code, so asset type 2 logged as a raw 0x02 and
journalctl hid the entire line as binary — concealing the field that identified
what was actually being refused.

+++ region/src/main.cpp
@@ -7327,8 +7327,13 @@ int main(int argc, char* argv[]) {
                                       << homeworldz::api::json_string(create_item->name)
                                       << ",\"wearableType\":"
                                       << static_cast<unsigned int>(create_item->wearable_type)
-                                      << ",\"assetType\":" << create_item->asset_type
-                                      << ",\"inventoryType\":" << create_item->inventory_type
+                                      // Cast: these are int8_t, and streaming
+                                      // one writes the character with that code
+                                      // rather than the number. Type 2 logged
+                                      // as a raw 0x02 and journalctl then hid
+                                      // the whole line as binary.
+                                      << ",\"assetType\":" << static_cast<int>(create_item->asset_type)
+                                      << ",\"inventoryType\":" << static_cast<int>(create_item->inventory_type)
                                       << ",\"wearable\":" << (wearable ? "true" : "false")
                                       << ",\"stagedAssetType\":" << staged_asset_type
                                       << ",\"assetIdResolved\":" << (asset_id.empty() ? "false" : "true")
